feat: move rules to dedicated subfolder
